Small Inv Change -> Andrew, please read description
Fully implemented skeleton of Inventory System. May make interface on
top of current abstract class to eliminate need for properties field.
Andrew: I had to go into Hero.cs? and modify the constructor that was
taking a new HeroAttackFactory because it was being passed a Stat item.
It was expecting a DungeonCharacter, so I changed it to
Hero.GetInstance(). Please check this out and change to what is
appropriate, I just patched it so code would compile, I have no idea of
that's what was needed. (I marked where the change happened in code.)